# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/website_sale_settings:
    get:
      tags:
        - WebsiteSaleSettings
      summary: List WebsiteSaleSettings
      description: Get a paginated list of WebsiteSaleSettings with optional filtering and sorting
      operationId: listWebsiteSaleSettings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteSaleSettingList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - WebsiteSaleSettings
      summary: Create WebsiteSaleSetting
      description: Create a new WebsiteSaleSetting
      operationId: createWebsiteSaleSetting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebsiteSaleSettingRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteSaleSetting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    WebsiteSaleSetting:
      type: object
      required:
        - id
        - website_id
        - access_gate
        - guest_party_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        access_gate:
          type: string
        default_customer_group_id:
          nullable: true
          type: string
          format: uuid
        guest_party_id:
          type: string
          format: uuid
        recovery_template_ref:
          nullable: true
          type: string
          maxLength: 200
        metadata:
          type: object
...
    CreateWebsiteSaleSettingRequest:
      type: object
      required:
        - website_id
        - access_gate
        - guest_party_id
        - metadata
      properties:
        website_id:
          type: string
          format: uuid
        access_gate:
          type: string
        default_customer_group_id:
          nullable: true
          type: string
          format: uuid
        guest_party_id:
          type: string
          format: uuid
        recovery_template_ref:
          nullable: true
          type: string
          maxLength: 200
        metadata:
          type: object
...
    WebsiteSaleSettingList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WebsiteSaleSetting'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []